Support nested query

+++ filter_generator.cpp
@@ -344,7 +344,7 @@ bson_t* Filter::generate_filter(std::string& field, std::string& term, std::stri
     size = tokens.size();
     filter = generate_unnested_filter(tokens.at(size - 1), term, dataType);
 
-    for (int i = size - 2; i >= 0; --i) {
+    for (long i = size - 2; i >= 0; --i) {
         filter = append_document(filter, tokens.at(i));
         std::cout << "index: " << i << std::endl;
     }
